Ensure a component is not smaller than its children
* Ensured that instantiated trees are have already been Laid out when returned
* Extracted layout delegates to separate file from layout rules
* Removed unnecessary calls to Layout() in tests
* Minor Work on TextInput and todos
