--- examples/todomvc/todomvc/components_test.go
@@ -1,22 +1,0 @@
-package todomvc
-
-import (
-	"assert"
-	"clock"
-	. "display"
-	"testing"
-)
-
-func TestCreate(t *testing.T) {
-
-	t.Run("Instantiable", func(t *testing.T) {
-		app, _ := Create(clock.NewFake())
-		assert.NotNil(t, app)
-	})
-
-	t.Run("Todo Component", func(t *testing.T) {
-		item, _ := Todo(NewBuilder(), Data(&TodoItemModel{Text: "abcd"}))
-		assert.Equal(t, item.Data().(*TodoItemModel).Text, "abcd")
-
-	})
-}
