+++ src/display/sparkline_test.go
@@ -7,7 +7,6 @@ func TestSparkline(t *testing.T) {
 		values := []float64{2.3, 11.3, 18.4, 13.5, 16.2, 63.1}
 		sparkline, _ := Sparkline(NewBuilder(), &SparklineModel{Values: values})
 
-		sparkline.Layout()
 		surface := NewFakeSurface()
 		sparkline.Draw(surface)
 	})
