+++ src/display/text_input_test.go
@@ -14,9 +14,13 @@ func TestTextInput(t *testing.T) {
 
 	t.Run("Placeholder text", func(t *testing.T) {
 		instance, _ := TextInput(NewBuilder(), Placeholder("Hello World"))
+		assert.Equal(t, instance.(*TextInputComponent).Placeholder(), "Hello World")
+	})
 
-		// Coerce to the concrete type.
-		textInput := instance.(*TextInputComponent)
-		assert.Equal(t, textInput.Placeholder(), "Hello World")
+	t.Run("Text() uses Placholder() when empty", func(t *testing.T) {
+		instance, _ := TextInput(NewBuilder(), Placeholder("abcd"))
+		assert.Equal(t, instance.Text(), "abcd")
+		instance.SetText("efgh")
+		assert.Equal(t, instance.Text(), "efgh")
 	})
 }
